docs : 메인 해태 말풍선 개별 컴포넌트 생성
