+++ src/pages/Main/Main.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useRef, useState } from "react";
-import gsap from "gsap";
 import { styled } from "styled-components";
 import NavBar from "../../component/NavBar/NavBar";
 import RightSide from "../../component/RightSide/RightSide";
@@ -8,14 +7,12 @@ import GiwaModal from "../../component/Modal/GiwaModal/GiwaModal";
 import Completed from "../../component/Popup/Completed";
 import BottomSide from "../../component/BottomSide/BottomSide";
 import GiwaButton from "../../component/GiwaButton/GiwaButton";
+import MainBg from "../../component/MainBg/MainBg";
 import mainHouse from "../../assets/main/giwa_house_indigo.png";
 import haetaeImg from "../../assets/main/haetae_img.png";
 import taegeukgi from "../../assets/main/taegeukgi.png";
-import pineTreeLeft from "../../assets/main/pine_tree_left.png";
-import pineTreeRight from "../../assets/main/pine_tree_right.png";
-import speechBubble from "../../assets/main/speech_bubble.svg";
-import { ReactComponent as GiwaPlus } from "../../assets/main/giwa_plus.svg";
 import Capture from "../../component/Popup/Capture";
+import Speech from "../../component/Speech/Speech";
 import { useBgColor } from "../../contexts/BackgroundColor"; // Bg Color Context
 
 const Main = () => {
@@ -25,8 +22,7 @@ const Main = () => {
   const [openMakeup, setOpenMakeup] = useState(false); // 기와집 꾸미기
   const [openGusetBook, setOpenGusetBook] = useState(false); // 방명록 모달창
   const [capturePopBol, setCapturePopBol] = useState(false); // 캡쳐 팝업
-  const [completedGiwa, setCompletedGiwa] = useState(false); // 기와 등록 팝업창
-  const speechRef = useRef();
+  const [completedGiwa, setCompletedGiwa] = useState(false); // 기와 등록 팝업창  
 
   /** 😀 juju
     - background useState는 하위 컴포넌트에 전역적으로 사용하기 위해...?
@@ -42,8 +38,6 @@ const Main = () => {
   useEffect(() => {
     // main페이지에서는 기와집을 불러오는 get요청을 해야함
     // 해당 api 확인되면 추가 예정
-
-    gsap.to(speechRef.current.querySelectorAll('span'), .5, { stagger: .1, display: 'inline', delay: .3, ease: 'Power1.easeInOut' })
   }, []);
 
   const openMakeupHouse = () => {
@@ -90,64 +84,14 @@ const Main = () => {
       <ExDiv $bgColor={bgColor}>
         <StyledMain>
           <HouseBox className={openMakeup || openGusetBook ? "left" : null}>
-            <HaetaeWrap>
-              <img src={haetaeImg} alt="해태" />
-              <Speech>
-                {/* {
-                  // 집주인 : 기와 12개 미만일 때
-                  <p ref={speechRef}>
-                    <span>환</span>
-                    <span>영</span>
-                    <span>하</span>
-                    <span>오</span>
-                    <span>!</span>
-                  </p>
-                } */}
-                {/* {
-                  // 집주인 : 기와 12개 이상일 때
-                  <>
-                    <p ref={speechRef}>
-                      <span>기</span>
-                      <span>와</span>
-                      <span>는</span>
-                      <span className="space">내</span>
-                      <span>가</span>
-                      <span className="space">보</span>
-                      <span>관</span>
-                      <span>하</span>
-                      <span>고</span>
-                      <span className="space">있</span>
-                      <span>소</span>
-                    </p>
-                    <button className="giwa_number">
-                      <span>10</span>
-                    </button>
-                  </>
-                } */}
-                { // 방문자
-                  <>
-                    <p ref={speechRef}>
-                      <span>기</span>
-                      <span>와</span>
-                      <span>를</span>
-                      <span className="space">남</span>
-                      <span>겨</span>
-                      <span>주</span>
-                      <span>시</span>
-                      <span>오</span>
-                      <span>!</span>
-                    </p>
-                    <button className="giwa_plus" onClick={() => setOpenModal(true)}>
-                      <GiwaPlus />
-                    </button>
-                  </>
-                }
-              </Speech>
-            </HaetaeWrap>
-            <img src={taegeukgi} alt="태극기" />
+            {/* 말풍선 start */}
+            <Speech setOpenModal={setOpenModal} />
+            {/* 말풍선 end */}
             {/* 기와 버튼 start */}
             <GiwaButton setOpen={openGusetBookModal} />
             {/* 기와 버튼 end */}
+            <img className="heatae" src={haetaeImg} alt="해태" />
+            <img className="taegeukgi" src={taegeukgi} alt="태극기" />
           </HouseBox>
         </StyledMain>
         <RightSide
@@ -167,25 +111,9 @@ const Main = () => {
           openMakeupHouse={openMakeupHouse}
           setCapturePopBol={setCapturePopBol}
         />
-        <Tree>
-          <img src={pineTreeLeft} alt="왼쪽 소나무" />
-          <img src={pineTreeRight} alt="오른쪽 소나무" />
-        </Tree>
-        {/* <Test>
-          {openMakeup || openGusetBook ? null : (
-            <button onClick={openMakeupHouse}>사용자 : 기와집 만들기</button>
-          )}
-          {openModal ? null : (
-            <button
-              onClick={() => {
-                setOpenModal(true);
-              }}
-              style={{ marginBottom: "50px" }}
-            >
-              방문자 : 기와선택
-            </button>
-          )}
-        </Test> */}
+        {/* 배경 start */}
+        <MainBg openMakeup={openMakeup} openGusetBook={openGusetBook} />
+        {/* 배경 end */}
         <Test2 onClick={changeBgColor}>밤/낮(toggle)</Test2>
       </ExDiv>
 
@@ -206,10 +134,10 @@ const ExDiv = styled.div`
   width: 100vw;
   height: 100vh;
   background: linear-gradient(
-    140deg,
+    158deg,
     ${({ $bgColor }) => $bgColor
     ? "#FFFEF9 0%, #FFF8DC 100%"
-    : " #8C92CA 0%, #31365B 100%"}
+    : " #868DCC 20%, #313557 95%"}
   );
   position: relative;
   overflow: hidden;
@@ -230,8 +158,8 @@ const HouseBox = styled.div`
   background: url(${mainHouse}) no-repeat;
   background-size: cover;
   position: absolute;
-  left: 7%;
-  top: 12%;
+  left: 120px;
+  top: 18%;
   right: 0;
   bottom: 0;
   margin: auto;
@@ -239,8 +167,14 @@ const HouseBox = styled.div`
   z-index: 2;
   > img {
     position: absolute;
-    left: -14%;
-    top: 18%;
+    &.heatae {
+      right: 37%;
+      top: 11.7%;
+    }
+    &.taegeukgi {
+      left: -14%;
+      top: 18%;
+    }
   }
   &.left {
     left: -500px;
@@ -256,82 +190,6 @@ const HaetaeWrap = styled.div`
   img {
     width: 100%;
     height: 100%;
-  }
-`;
-
-const Speech = styled.div`
-  width: 242px;
-  height: 64px;
-  position: absolute;
-  top: -41%;
-  left: -26%;
-  margin: auto;
-  background: url(${speechBubble}) 50% 50% no-repeat;
-  background-size: cover;
-  text-align: center;
-  > button {
-    &.giwa_plus {
-      position: absolute; 
-      top: -35%;
-      right: 15%;
-      margin: auto; 
-    }
-    &.giwa_number {
-      position: absolute;
-      padding: 0 14px 3px;
-      background-color: #6C5847;
-      border-radius: 50px;
-      right: 10%;
-      top: -15%;
-      span {
-        color: #fff;
-        font-family: var(--font-hunmin);
-        font-size: 12px;
-        font-weight: 400;
-        line-height: 16px;
-        letter-spacing: 0.12px;
-      }
-    }
-  }
-  > p { 
-    margin: 9% 0 0;
-    display: inline-block; 
-    color: #222;
-    font-family: var(--font-hunmin);
-    font-size: 16px;
-    font-weight: 600;
-    letter-spacing: 0.16px;
-    span {
-      display: none;
-      &.space {
-        margin: 0 0 0 5px;
-      }
-    }
-  }
-`;
-
-
-
-const Tree = styled.div`
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  left: 0;
-  top: 0;
-  pointer-events: none;
-  > img {    
-    position: absolute;
-    &:nth-of-type(1) {
-      max-width: 460px;
-      left: 0;
-      bottom: 0;
-      z-index: 2;
-    }
-    &:nth-of-type(2) {
-      max-width: 410px;
-      right: 0;
-      top: 13%;
-    }
   }
 `;
